+++ Sources/MapboxMaps/ContentBuilders/MapContent/ForEvery.swift
@@ -25,6 +25,7 @@
 /// }
 /// ```
 ///
+///
 /// - Note: `ForEvery` is similar to SwiftUI `ForEach`, but works with ``MapContent``.
 @available(iOS 13.0, *)
 public struct ForEvery<Content, Data: RandomAccessCollection, ID: Hashable> {
@@ -54,6 +55,9 @@ public struct ForEvery<Content, Data: RandomAccessCollection, ID: Hashable> {
 @available(iOS 13.0, *)
 extension ForEvery: MapContent, PrimitiveMapContent where Content: MapContent {
     /// Creates instance that identified data by given key path.
+    ///
+    /// It’s important that the id of a data element doesn’t change, unless the data element has been replaced with a new data element that has a new identity.
+    /// If two elements with the same id are passed, the behavior not guaranteed.
     public init(_ data: Data, id: KeyPath<Data.Element, ID>, @MapContentBuilder content: @escaping (Data.Element) -> Content) {
         self.init(data: data, id: id, content: content)
     }